✨ Add slack and giphy services
