+++ .env.example
@@ -1,0 +1,2 @@
+SLACK_TOKEN=
+GIPHY_API_KEY=
