+++ package.json
@@ -8,7 +8,10 @@
     "test:notify": "tsx src/test-notify.ts"
   },
   "dependencies": {
+    "@slack/web-api": "^7.3.1",
+    "axios": "^1.7.2",
     "chalk": "^5.3.0",
+    "dotenv": "^16.4.5",
     "node-notifier": "^10.0.0",
     "tsx": "^4.16.2"
   },
