+++ src/index.ts
@@ -1,11 +1,7 @@
 import { ERROR_MSG_PLAYERS } from "./constants"
-import { logger } from "./logger"
-import {
-  getRoles,
-  getSessionDetailsFromUserInput,
-  isValidPlayerCount,
-  rotatePlayers,
-} from "./util"
+import { logger } from "./utils/logger"
+import { isValidPlayerCount, rotatePlayers } from "./utils/rotation"
+import { getSessionDetailsFromUserInput } from "./utils/user-input"
 
 const main = async () => {
   logger.info("Starting Mob Programming Session...")
